Move : 패키지 분리
Domain, Global 로 패키지 분리
