+++ src/main/java/org/example/bikers/domain/bikeModel/entity/Manufacturer.java
@@ -1,4 +1,4 @@
-package org.example.bikers.bikeModel.entity;
+package org.example.bikers.domain.bikeModel.entity;
 
 public enum Manufacturer {
     HONDA,
