+++ src/main/java/org/example/bikers/domain/bikeModel/repository/BikeModelRepository.java
@@ -1,6 +1,6 @@
-package org.example.bikers.bikeModel.repository;
+package org.example.bikers.domain.bikeModel.repository;
 
-import org.example.bikers.bikeModel.entity.BikeModel;
+import org.example.bikers.domain.bikeModel.entity.BikeModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BikeModelRepository extends JpaRepository<BikeModel, Long> {
