+++ src/main/java/org/example/bikers/domain/bikeModel/service/BikeModelService.java
@@ -1,4 +1,4 @@
-package org.example.bikers.bikeModel.service;
+package org.example.bikers.domain.bikeModel.service;
 
 import org.springframework.stereotype.Service;
