+++ src/main/java/org/example/bikers/domain/member/dto/MemberSignupRequestDto.java
@@ -1,4 +1,4 @@
-package org.example.bikers.member.dto;
+package org.example.bikers.domain.member.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
