+++ src/main/java/org/example/bikers/domain/member/entity/MemberStatus.java
@@ -1,4 +1,4 @@
-package org.example.bikers.member.entity;
+package org.example.bikers.domain.member.entity;
 
 public enum MemberStatus {
     NORMAL,
