+++ src/main/java/org/example/bikers/domain/member/repository/MemberRepository.java
@@ -1,7 +1,7 @@
-package org.example.bikers.member.repository;
+package org.example.bikers.domain.member.repository;
 
 import java.util.Optional;
-import org.example.bikers.member.entity.Member;
+import org.example.bikers.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
