+++ src/main/java/org/example/bikers/domain/myBike/controller/MyBikeController.java
@@ -1,4 +1,4 @@
-package org.example.bikers.myBike.controller;
+package org.example.bikers.domain.myBike.controller;
 
 import org.springframework.web.bind.annotation.RestController;
