+++ src/main/java/org/example/bikers/domain/myBike/repository/MyBikeRepository.java
@@ -1,6 +1,6 @@
-package org.example.bikers.myBike.repository;
+package org.example.bikers.domain.myBike.repository;
 
-import org.example.bikers.myBike.entity.MyBike;
+import org.example.bikers.domain.myBike.entity.MyBike;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MyBikeRepository extends JpaRepository<MyBike, Long> {
